// apps/web/src/lib/api/generated/model/invoiceResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Petclinic MSS301 — Aggregated API
 * Aggregated from: auth, customers, vets, visits, workflow, billing, products, reviews
 * OpenAPI spec version: 1.0.0
 */
import type { InvoiceResponseStatus } from './invoiceResponseStatus';
import type { InvoiceResponsePaymentMethod } from './invoiceResponsePaymentMethod';
import type { InvoiceItemResponse } from './invoiceItemResponse';

export interface InvoiceResponse {
  id?: number;
  customerUserId?: string;
  customerName?: string;
  customerEmail?: string;
  status?: InvoiceResponseStatus;
  currency?: string;
  subtotal?: number;
  total?: number;
  notes?: string;
  issuedAt?: string;
  paidAt?: string;
  paymentMethod?: InvoiceResponsePaymentMethod;
  paymentReference?: string;
  items?: InvoiceItemResponse[];
}